Edit schemas

+++ commerce/shopping-store/src/main/resources/schema.sql
@@ -9,5 +9,5 @@ CREATE TABLE IF NOT EXISTS products (
     quantity_state    VARCHAR(50) NOT NULL,
     product_state     VARCHAR(50) NOT NULL,
     product_category  VARCHAR(50) NOT NULL,
-    price             FLOAT NOT NULL
+    price             NUMERIC(10, 2) NOT NULL
 );
